requirements.txt
imageio
pillow
pandas
numpy
scikit-learn
